Propagate syntax error from server side rendering require
This will propagate syntax errors from the require statement to the
error prop in the loading component.

The biggest issue with silently catching these errors is that files that
have special loaders in webpack will fail with syntax error and might be
hard to track down since they differ from the client side environment.